requirements.txt
Flask
python-dotenv
openai
gunicorn
requests
asyncio
Flask[async]

gunicorn
uvicorn
uvicorn[standard]
fastapi